H.Res. 61 Reported in House
Providing for consideration of the bill (H.R. 648) making appropriations for the fiscal year ending September 30, 2019, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 31) making further continuing appropriations for the Department of Homeland Security for fiscal year 2019, and for other purposes; providing for consideration of motions to suspend the rules; and waiving a requirement of clause 6(a) of rule XIII with respect to consideration of certain resolutions reported from the Committee on Rules.

Version:  Reported in House (rh)
Date:     2019-01-22
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres61rh/xml/BILLS-116hres61rh.xml

Sponsored-By: M001206