H.Res. 61 Engrossed in House
Providing for consideration of the bill (H.R. 648) making appropriations for the fiscal year ending September 30, 2019, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 31) making further continuing appropriations for the Department of Homeland Security for fiscal year 2019, and for other purposes; providing for consideration of motions to suspend the rules; and waiving a requirement of clause 6(a) of rule XIII with respect to consideration of certain resolutions reported from the Committee on Rules.

Version:  Engrossed in House (eh)
Date:     2019-01-23
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres61eh/xml/BILLS-116hres61eh.xml

Sponsored-By: M001206
Roll-Call: House 116-1-47 2019-01-23 Passed 223-190