H.Res. 105 Engrossed in House
Providing for consideration of the bill (H.R. 840) to amend title 38, United States Code, to direct the Secretary of Veterans Affairs to provide child care assistance to veterans receiving certain medical services provided by the Department of Veterans Affairs; providing for the adoption of the resolution (H. Res. 86) providing amounts for the expenses of the Select Committee on the Climate Crisis and the Select Committee on the Modernization of Congress; and providing for consideration of motions to suspend the rules.

Version:  Engrossed in House (eh)
Date:     2019-02-07
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres105eh/xml/BILLS-116hres105eh.xml

Sponsored-By: M001206