H.Res. 252 Engrossed in House
Providing for consideration of the bill (H.R. 7) to amend the Fair Labor Standards Act of 1938 to provide more effective remedies to victims of discrimination in the payment of wages on the basis of sex, and for other purposes, and providing for consideration of the resolution (H. Res. 124) expressing opposition to banning service in the Armed Forces by openly transgender individuals.

Version:  Engrossed in House (eh)
Date:     2019-03-27
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres252eh/xml/BILLS-116hres252eh.xml

Sponsored-By: T000474