S.Res. 96 Agreed to Senate
A resolution commending the Government of Canada for upholding the rule of law and expressing concern over actions by the Government of the People's Republic of China in response to a request from the United States Government to the Government of Canada for the extradition of a Huawei Technologies Co., Ltd. executive.

Version:  Agreed to Senate (ats)
Date:     2019-05-07
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116sres96ats/xml/BILLS-116sres96ats.xml

Sponsored-By: R000584
Cosponsor-Count: 10