H.Res. 431 Reported in House
Providing for consideration of the bill (H.R. 2740) making appropriations for the Departments of Labor, Health and Human Services, and Education, and related agencies for the fiscal year ending September 30, 2020, and for other purposes, and providing for consideration of the resolution (H. Res. 430) authorizing the Committee on the Judiciary to initiate or intervene in judicial proceedings to enforce certain subpoenas and for other purposes.

Version:  Reported in House (rh)
Date:     2019-06-10
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres431rh/xml/BILLS-116hres431rh.xml

Sponsored-By: R000606