H.Res. 553 Engrossed in House
Directing the Clerk of the House of Representatives to provide a copy of the on-the-record portions of the audio backup file of the transcribed interview of Roger J. Stone Jr. conducted by the Permanent Select Committee on Intelligence on September 26, 2017, to the prosecuting attorneys in the case of United States of America v. Stone, No. 1:19-cr-00018-ABJ (D.D.C.).

Version:  Engrossed in House (eh)
Date:     2019-09-12
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres553eh/xml/BILLS-116hres553eh.xml

Sponsored-By: M000312