H.Res. 577 Reported in House
Providing for consideration of the bill (H.R. 2203) to increase transparency, accountability, and community engagement within the Department of Homeland Security, provide independent oversight of border security activities, improve training for agents and officers of U.S. Customs and Border Protection and U.S. Immigration and Customs Enforcement, and for other purposes; providing for consideration of the bill (H.R. 3525) to amend the Homeland Security Act of 2002 to direct the Commissioner of U.S. Customs and Border Protection to establish uniform processes for medical screening of individuals interdicted between ports of entry, and for other purposes; providing for consideration of the resolution (H. Res. 576) expressing the sense of the House of Representatives with respect to the whistleblower complaint of August 12, 2019, made to the Inspector General of the Intelligence Community; and for other purposes.

Version:  Reported in House (rh)
Date:     2019-09-24
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres577rh/xml/BILLS-116hres577rh.xml

Sponsored-By: S001205