H.Res. 741 Engrossed in House
Providing for consideration of the bill (H.R. 4) to amend the Voting Rights Act of 1965 to revise the criteria for determining which States and political subdivisions are subject to section 4 of the Act, and for other purposes, and providing for consideration of the resolution (H. Res. 326) expressing the sense of the House of Representatives regarding United States efforts to resolve the Israeli-Palestinian conflict through a negotiated two-state solution.

Version:  Engrossed in House (eh)
Date:     2019-12-05
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres741eh/xml/BILLS-116hres741eh.xml

Sponsored-By: R000606