H.Res. 765 Engrossed in House
Providing for consideration of the Senate amendment to the bill (H.R. 1158) to authorize cyber incident response teams at the Department of Homeland Security, and for other purposes; providing for consideration of the Senate amendment to the bill (H.R. 1865) to require the Secretary of the Treasury to mint a coin in commemoration of the opening of the National Law Enforcement Museum in the District of Columbia, and for other purposes; and providing for the adoption of the resolution (H. Res. 761) permitting individuals to be admitted to the Hall of the House in order to obtain footage of the House in session for inclusion in the orientation film to be shown to visitors at the Capitol Visitor Center.

Version:  Engrossed in House (eh)
Date:     2019-12-17
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres765eh/xml/BILLS-116hres765eh.xml

Sponsored-By: M001206
Roll-Call: House 116-1-683 2019-12-17 Passed 220-192
Roll-Call: House 116-1-684 2019-12-17 Passed 219-189