H.Res. 844 Engrossed in House
Providing for consideration of the bill (H.R. 2546) to designate certain lands in the State of Colorado as components of the National Wilderness Preservation System, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 79) removing the deadline for the ratification of the equal rights amendment; and for other purposes.

Version:  Engrossed in House (eh)
Date:     2020-02-11
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres844eh/xml/BILLS-116hres844eh.xml

Sponsored-By: S001205