H.Res. 1017 Engrossed in House
Providing for consideration of the bill (H.R. 51) to provide for the admission of the State of Washington, D.C. into the Union; providing for consideration of the bill (H.R. 1425) to amend the Patient Protection and Affordable Care Act to provide for a Improve Health Insurance Affordability Fund to provide for certain reinsurance payments to lower premiums in the individual health insurance market; providing for consideration of the bill (H.R. 5332) to amend the Fair Credit Reporting Act to ensure that consumer reporting agencies are providing fair and accurate information reporting in consumer reports, and for other purposes; providing for consideration of the bill (H.R. 7120) to hold law enforcement accountable for misconduct in court, improve transparency through data collection, and reform police training and policies; providing for consideration of the bill (H.R. 7301) to prevent evictions, foreclosures, and unsafe housing conditions resulting from the COVID-19 pandemic, and for other purposes; providing for consideration of the joint resolution (H.J. Res. 90) providing for congressional disapproval under chapter 8 of title 5, United States Code, of the rule submitted by the Office of the Comptroller of the Currency relating to Community Reinvestment Act Regulations; and for other purposes.

Version:  Engrossed in House (eh)
Date:     2020-06-25
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres1017eh/xml/BILLS-116hres1017eh.xml

Sponsored-By: H000324