H.Res. 1107 Reported in House
Providing for consideration of the bill (H.R. 2574) to amend title VI of the Civil Rights Act of 1964 to restore the right to individual civil actions in cases involving disparate impact, and for other purposes; providing for consideration of the bill (H.R. 2639) to establish the Strength in Diversity Program, and for other purposes; providing for consideration of the bill (H.R. 2694) to eliminate discrimination and promote women’s health and economic security by ensuring reasonable workplace accommodations for workers whose ability to perform the functions of a job are limited by pregnancy, childbirth, or a related medical condition; providing for consideration of the resolution (H. Res. 908) condemning all forms of anti-Asian sentiment as related to COVID–19; and for other purposes.

Version:  Reported in House (rh)
Date:     2020-09-14
Congress: 116

Source: https://www.govinfo.gov/content/pkg/BILLS-116hres1107rh/xml/BILLS-116hres1107rh.xml

Sponsored-By: S001205